dnsdist: Split packet processing functions out of `dnsdist.cc`
